backend/python/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
asyncpg==0.29.0
sqlalchemy==2.0.23
alembic==1.12.1

# BigQuery
google-cloud-bigquery==3.13.0
google-cloud-bigquery-storage==2.22.0
pandas-gbq==0.19.2

# Analytics & ML
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
tensorflow==2.15.0

# Natural Language Processing
openai==1.3.7
langchain==0.0.340
transformers==4.35.2
nltk==3.8.1

# Redis
redis==5.0.1
aioredis==2.0.1

# Utils
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0
httpx==0.25.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1
ipython==8.18.1